pyproject.toml
[project]
name = "adk-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.24.1",
    "google-cloud-aiplatform>=1.135.0",
    "requests>=2.31.0",
    "python-dotenv>=1.2.1",
    "Pillow>=10.0.0",
    "langchain_google_vertexai>=3.2.2",
    "langgraph>=1.0.7",
    "opentelemetry-instrumentation-google-genai",
    "google-genai>=1.56.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["lark_agent*"]